students/YarmolovichAleksandr/task_02/src/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-restx
Flask-Marshmallow
marshmallow
marshmallow-sqlalchemy
python-dotenv
flask-cors
pytest
pytest-flask
psycopg2-binary
bcrypt
python-multipart